adapt "sed" to work on mac
freebsd seems to be different from unix here.
i did not figure out how to adapt the statement
that adds the "GENERATED" hint to the content.
however, comments in markdown are not really standard anyway,
so i decided to make the hint in the README clearer.

in general, i think the t-dance script
should be replaced by a python script as some point -
but, for now, it works again ;)