hotfix: force invalidating javascript cache
with the last update of the download-page,
the name of the element "osx" as changes to "macos" -
this was also changed in assets/js/download-page.js,
however, depending on the browser, cached files are not invalidated.

this might be fixed by some server settings, however,
after all experience with that, it is easier and also more reliable
to just news a new name.