add information about minimal requirements
i think the "Changelog" section is the closest match and not that wrong,
adding it directly below the boxes looks bad
and i do not want to add an extra headline,
there is no need to make this too much outstanding, ppl will find it.
for the same reason, i do not want to add it directly to the boxes.