alternative blog preview
due to a lack of a paging functionality
(which i personally do not miss ;)
and to keep the overall page size small,
we did not add an image for the 10th oldest blob post.

this results in some "text desert"
that is not really nice and makes the page
a bit unclear when scolling down.
too many information :)

this commit changes that by just adding the clickable title
for older posts.
moreover, the "new posts" are increased to 12 (so, roughly a year)

i find the blog much nicer this way,
see preview.