split first paragraph of ios-blogpost
the first paragraph is shown in the blog overview
and is a bit long there, also the "Note ..."
sentence does not make much sense in the overview.

with this change, the "Read on..." link will appear one sentence before,
the remaining first paragraph has a reasonable length,
similar to all the other posts then.