unify changelog urls
core/desktop/ios point to the normal github repo
that shows the changelog rich formatted, including github navigation.

android point to the raw, unformatted textfile.

boths may have advantages and disadvantages,
however, i do not know of reasons to mix the approaches.

so, this pr changes android to do the same as the other repos.

(the include file should stay as is until translations are done
and the include file is not used anywhere)