avoid links in headlines that go to TOC
these links look weird in the TOC - but also as a real headline.

also, it feels wrong to have a first link that tend frighten ppl -
where only the next link to our blog says, "no, everything is fine".

so, the link to the blog post, is the better source for
diving deeper into the topic, as it deals with
Delta Chat wrt EFAIL and not only EFAIL.

finally, we also do not link PGP, OpenPGP etc. in the headlines
(a link is not really a part of a question).