use unagitated, more professional looking list
while on making the page easier translatable
by keep using markdown and increase the fontsize by some css magic,
it turns out that the simple thing looks better.

the increased fontsize
making headlines less outstanding and turns the whole page
a bit into a desert of text without structure.
also, it does not look good compared the rest of the site.

i played around with some background or a box - however,
there is not much on the page,
why adding another level of attracting attention.

so it seems better to stay with the simple thing
and better concentrate making the screenshots nice.