recent android update will require android5
the bump was needed to target api34,
androidx dropped support for android4.
this happend already before for other libs and for the ndk (!),
and we always found workarounds sooner or later,
but this time, _many_ workarounds would be needed,
this is no longer reasonable.

5 years ago, we were already considering to do that step,
that time affecting 5% of users ... https://github.com/deltachat/deltachat-android/pull/1115

so, one can say we gave these users 5 more years a chance,
and 90% of them meanwhile got a new device or did an update.