show Github download option
this is useful, in case delta.chat is unavailable for whatever reason.

with LibResilient it may be possible that the blocked page is still openable.
but even if not, it is useful to show the option,
so that ppl get to know this option.

(we're not linking to the concrete release version,
however, as that may be out of sync with the pages-version -
and we do not want to add additional maintainance effort here)