play around with client-side redirecs
server-side redirects cause a lot of surprisingly hard to fix issues -
as the years-old issue with the /en/en/en/en/... redirects (#848),
but also LibResilient has issues with them.

surely, this is all solvable,
however, another solution are good old client-wise redirects.

this means, to have a little html-file for each redirect
(as dl.html in this example),
the file contains the destination (relatively!) as HTML,
and also executes a little javascript that

drawbacks are:
- when javascript is disabled, things are in english
- while search engined follow client-side-redirects,
  they may rank pages using server-side redirects better.

the drawbacks seem not to outweight the issues before
nor affect a significant number of users or have big imact.