fix blog links in mirrors
mirrors may be placed in a subdirectory,
this way, links continue to work;
we use the `../` approach also elsewhere, eg. on language selection,
so this does not introduce smth new.

baseurl seems just not to be set,
and it is also not really used elsewhere.

may still be nice to hace mirrors on root,
however, with this PR this is less a prerequise.