"add second device" works with clipboard as well
briefly mention the clipboard option,
it is complex, if one does not have an idea how clipboard etc. works,
ppl anyway need help.

doing transfer over clipboard is preferrable to transferring backup files -
it is easier, and usually more secure.

it is still fine to have that at the pretty end,
other issues are still more important,
surely not all towers have a cam, but many do, and most notebooks as well.
the bias in our circles may be different, of course :)